Correct naming conventions for site identifiers
Unlike stated in the documentation, the `.` as part of the site identifier only partially works. It is fine as long as you generate the folder manually, but once you edit the site configuration with that site identifier in the backend, the `.` will be removed and a second folder, excluding the dots will be generated.